Move AbstractSpecimenTask, FileAnalysisSpecimenTask, StandaloneSpecimenTask, and StudyImportSpecimenTask to study-api
