+++ study/api-src/org/labkey/api/specimen/pipeline/FileAnalysisSpecimenTask.java
@@ -1,4 +1,4 @@
-package org.labkey.study.pipeline;
+package org.labkey.api.specimen.pipeline;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.BooleanUtils;
@@ -8,16 +8,13 @@ import org.apache.tika.metadata.Metadata;
 import org.apache.tika.mime.MediaType;
 import org.jetbrains.annotations.Nullable;
 import org.labkey.api.admin.ImportException;
-import org.labkey.api.admin.PipelineJobLoggerGetter;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.pipeline.PipelineJobException;
 import org.labkey.api.pipeline.file.FileAnalysisJobSupport;
-import org.labkey.api.specimen.pipeline.AbstractSpecimenTaskFactory;
 import org.labkey.api.study.importer.SimpleStudyImportContext;
 import org.labkey.api.util.DateUtil;
 import org.labkey.api.writer.FileSystemFile;
 import org.labkey.api.writer.VirtualFile;
-import org.labkey.study.importer.StudyImportContext;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -34,7 +31,7 @@ public class FileAnalysisSpecimenTask extends AbstractSpecimenTask<FileAnalysisS
     }
 
     @Override
-    protected File getSpecimenFile(PipelineJob job) throws Exception
+    protected File getSpecimenFile(PipelineJob job)
     {
         FileAnalysisJobSupport support = job.getJobSupport(FileAnalysisJobSupport.class);
 
@@ -43,11 +40,12 @@ public class FileAnalysisSpecimenTask extends AbstractSpecimenTask<FileAnalysisS
         return support.getInputFiles().get(0);
     }
 
-    @Override
-    SimpleStudyImportContext getImportContext(PipelineJob job)
-    {
-        return new StudyImportContext(job.getUser(), job.getContainer(), null, new PipelineJobLoggerGetter(job));
-    }
+//  I think this can be deleted... just rely on impl in AbstractSpecimenTask
+//    @Override
+//    SimpleStudyImportContext getImportContext(PipelineJob job)
+//    {
+//        return new StudyImportContext(job.getUser(), job.getContainer(), null, new PipelineJobLoggerGetter(job));
+//    }
 
     @Override
     protected boolean isMerge()
